// src/interfaces/employee-details/employee-details.ts
export interface EmployeeSelfDetailsType {
  employeeProfileLandingView: EmployeeProfileLandingView;
  userVM: UserVM;
  empEmployeeAddress?: EmpEmployeeAddressEntity[] | null;
  empEmployeeBankDetail: EmpEmployeeBankDetail;
  empEmployeeEducation?: null[] | null;
  empEmployeeRelativesContact?: EmpEmployeeRelativesContactEntity[] | null;
  empEmployeeJobHistory?: null[] | null;
  empEmployeeTraining?: null[] | null;
  empEmployeePhotoIdentity: EmpEmployeePhotoIdentity;
  empJobExperience?: null[] | null;
  empSocialMedia?: null[] | null;
}
export interface EmployeeProfileLandingView {
  intEmployeeBasicInfoId: number;
  strEmployeeName: string;
  strEmployeeCode: string;
  strReferenceId: string;
  intEmployeeImageUrlId: number;
  intDesignationId: number;
  strDesignation: string;
  intDepartmentId: number;
  strDepartment: string;
  intSupervisorId: number;
  strSupervisorName: string;
  intSupervisorImageUrlId: number;
  intLineManagerId: number;
  strLinemanager: string;
  intLinemanagerImageUrlId: number;
  strCardNumber: string;
  intGenderId: number;
  strGender: string;
  intReligionId: number;
  strReligion: string;
  strMaritalStatus: string;
  strBloodGroup: string;
  dteDateOfBirth: string;
  dteJoiningDate: string;
  dteInternCloseDate?: null;
  dteProbationaryCloseDate?: null;
  dteContractFromDate?: null;
  dteContractToDate?: null;
  dteConfirmationDate?: null;
  dteLastWorkingDate?: null;
  intDottedSupervisorId: number;
  strDottedSupervisorName: string;
  strServiceLength: string;
  isSalaryHold: boolean;
  isActive: boolean;
  isUserInactive: boolean;
  isRemoteAttendance: boolean;
  intWorkplaceId: number;
  strWorkplaceName: string;
  intWorkplaceGroupId: number;
  strWorkplaceGroupName: string;
  intBusinessUnitId: number;
  strBusinessUnitName: string;
  intAccountId: number;
  strAccountName: string;
  strBankAccountName?: null;
  strBankAccountNo?: null;
  strRoutingNo?: null;
  strBranchName?: null;
  strBankWalletName?: null;
  dteCreatedAt: string;
  intCreatedBy: number;
  dteUpdatedAt?: null;
  intUpdatedBy: number;
  intEmploymentTypeId: number;
  strEmploymentType: string;
  employmentType?: null;
  intDetailsId: number;
  strOfficeMail: string;
  strPersonalMail: string;
  strPersonalMobile: string;
  strOfficeMobile: string;
  intPayrollGroupId: number;
  strPayrollGroupName: string;
  intPayscaleGradeId: number;
  strPayscaleGradeName: string;
  intCalenderTypeId: number;
  strCalenderType: string;
  intCalenderId: number;
  strCalenderName: string;
  intHrpositionId: number;
  strHrpostionName: string;
  intEmployeeStatusId: number;
  strEmployeeStatus: string;
  sectionId: number;
  sectionName: string;
  gradeId: number;
  gradeName: string;
  categoryName: string;
  intParentId: number;
  isManual: number;
  strLoginId?: null;
  intCountryId?: null;
  strCountry?: null;
  intUserTypeId?: null;
  strPassword?: null;
  strPersonalEmail?: null;
  intUserId?: null;
  isCreateUser: boolean;
  strUserType?: null;
  userStatus?: null;
  isTakeHomePay: boolean;
}
export interface UserVM {
  loginId: string;
  strPassword: string;
  officeMail: string;
  strPersonalMobile: string;
  userTypeId: number;
  strUserType: string;
  userStatus: boolean;
}
export interface EmpEmployeeAddressEntity {
  intEmployeeAddressId: number;
  intEmployeeBasicInfoId: number;
  intAddressTypeId: number;
  strAddressType: string;
  intCountryId: number;
  strCountry: string;
  intDivisionId: number;
  strDivision: string;
  strDistrictOrStateId: number;
  strDistrictOrState: string;
  strAddressDetails: string;
  intPostOfficeId?: null;
  strPostOffice?: null;
  strZipOrPostCode: string;
  isActive: boolean;
  dteCreatedAt: string;
  intCreatedBy: number;
  dteUpdatedAt?: null;
  intUpdatedBy?: null;
}
export interface EmpEmployeeBankDetail {
  intEmployeeBankDetailsId: number;
  intEmployeeBasicInfoId: number;
  intBankOrWalletType: number;
  intBankWalletId: number;
  strBankWalletName: string;
  strDistrict: string;
  intBankBranchId: number;
  strBranchName: string;
  strRoutingNo: string;
  strSwiftCode: string;
  strAccountName: string;
  strAccountNo: string;
  isPrimarySalaryAccount: boolean;
  isActive: boolean;
  intWorkplaceId: number;
  intBusinessUnitId: number;
  intAccountId: number;
  dteCreatedAt: string;
  intCreatedBy: number;
  dteUpdatedAt: string;
  intUpdatedBy: number;
}
export interface EmpEmployeeRelativesContactEntity {
  intEmployeeRelativesContactId: number;
  intEmployeeBasicInfoId: number;
  strRelativesName: string;
  intRelationShipId: number;
  strRelationship: string;
  strPhone: string;
  strEmail: string;
  strAddress: string;
  isEmergencyContact: boolean;
  intGrantorNomineeTypeId: number;
  strGrantorNomineeType: string;
  strNid: string;
  strBirthId: string;
  dteDateOfBirth: string;
  strRemarks: string;
  intPictureFileUrlId: number;
  isActive: boolean;
  dteCreatedAt: string;
  intCreatedBy: number;
  dteUpdatedAt: string;
  intUpdatedBy: number;
}
export interface EmpEmployeePhotoIdentity {
  intEmployeePhotoIdentityId: number;
  intEmployeeBasicInfoId: number;
  intProfilePicFileUrlId: number;
  intProfilePicFormalFileUrlId?: null;
  intSignatureFileUrlId?: null;
  strNid?: null;
  intNidfileUrlId?: null;
  strBirthId?: null;
  intBirthIdfileUrlId?: null;
  strPassport?: null;
  intPassportFileUrlId?: null;
  strNationality?: null;
  strBiography?: null;
  strHobbies?: null;
  isActive: boolean;
  dteCreatedAt: string;
  intCreatedBy: number;
  dteUpdatedAt?: null;
  intUpdatedBy?: null;
}